[libs] E: Move Binding for getsockname()
This commits moves the binding for the getsockname() system calls to the
syscall crate.
